requirements.txt
fastapi
uvicorn[standard]
websockets
robin-stocks
yfinance
python-dotenv
pandas
requests
pyjwt
